+++ gallery/charts-gallery/js/main.js
@@ -51,6 +51,9 @@ $(window).on("load", function () {
             fillColors: '#ffecad',
             negativeFillColors: '#ffb0b0',
             strokeColors: '#9c9c9c',
+            labelsY: {
+                captions: "=v/1000 + 'k'",
+            },
             colHeightStart: 0.1,
             props: [{
                     prop: 'col-height',
@@ -59,8 +62,8 @@ $(window).on("load", function () {
         }, {
             chartType: ABeamer.ChartTypes.area,
             data: [
-                [30000, 42000, 50000, 30000, 45000, 40000, 30000],
-                [13000, 32000, 35000, 55000, 35000, 50000, 25000],
+                [30, 42, 50, 30, 45, 40, 30],
+                [13, 32, 35, 55, 35, 50, 25],
             ],
             colInterSpacing: 1,
             colSpacing: 20,
@@ -68,6 +71,7 @@ $(window).on("load", function () {
             labelsX: {
                 captions: ['Q1', 'Q2', 'Q3', 'Q4', '18-Q1', 'Q2', 'Q3'],
             },
+            labelsY: "=v + '%'",
             fillColors: ['#35a9c070', '#adffb970'],
             strokeColors: ['#9c9c9c', '#9c9c9c'],
             strokeWidth: [2, 2],
@@ -108,6 +112,7 @@ $(window).on("load", function () {
                     prop: 'col-height',
                     value: 1,
                 }],
+            chartWidth: 320,
         }, {
             chartType: ABeamer.ChartTypes.bar,
             labelsX: defLabelsX,
@@ -151,6 +156,10 @@ $(window).on("load", function () {
     //                               Factory
     // ------------------------------------------------------------------------
     var chartCount = !firstOnly ? charts.length : 1;
+    var widths = [];
+    charts.forEach(function (chart, i) {
+        widths.push(chart.chartWidth || 280);
+    });
     scene1
         .addAnimations([{
             selector: '#gallery',